requirements.txt
asgiref==3.4.1
astroid==2.8.0
autopep8==1.5.7
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.1
cloudinary==1.28.0
colorama==0.4.4
cryptography==36.0.0
defusedxml==0.7.1
dj-static==0.0.6
Django==3.2.7
django-allauth==0.46.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mercadopago==2.0.8
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.12.1
